+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Please make sure to add your changes to the appropriate categories:
 ### Changed
 
 - Bumped MSRV from `1.70.0` to `1.72.0`.
+- Replaced panics with `Result` returns in control methods.
 
 ### Deprecated